+++ passes/V008/V008_test.go
@@ -3,7 +3,6 @@ package V008_test
 import (
 	"testing"
 
-	"github.com/bflad/tfproviderlint/helper/analysisfixtest"
 	"github.com/bflad/tfproviderlint/passes/V008"
 	"golang.org/x/tools/go/analysis/analysistest"
 )
@@ -15,5 +14,5 @@ func TestV008(t *testing.T) {
 
 func TestAnalyzerFixes(t *testing.T) {
 	testdata := analysistest.TestData()
-	analysisfixtest.Run(t, testdata, V008.Analyzer, "a")
+	analysistest.RunWithSuggestedFixes(t, testdata, V008.Analyzer, "a")
 }